% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dh.test.R
\name{dh.test}
\alias{dh.test}
\title{Normality and White Noise Testing}
\usage{
dh.test(x)
}
\arguments{
\item{x}{is a vector.}
}
\value{
The p-value for the test.
}
\description{
Normality and White Noise Testing
}
\details{
This function is a modification of the normwhn.test::normality.test1 function
}
\examples{
x <- rnorm(n=30, mean=0, sd=1)
dh.test(x)
}
